make OSCPortInBuilder a separate (non-nested) class; formatting fixes
